feat: [deadcode-cleanup-003] - [Remove confirmed dead backend code without runtime behavior changes]
