+++ tests/functional/bootstrap_portability/token_helpers_test.go
@@ -1,5 +1,3 @@
-//go:build functionallong
-
 package bootstrap_portability
 
 import (
