+++ tests/functional/internal/support/events.go
@@ -149,16 +149,6 @@ func appendUniqueDispatchWorkID(values []string, value string) []string {
 	return append(values, value)
 }
 
-func LastFactoryEventTick(events []factoryapi.FactoryEvent) int {
-	tick := 0
-	for _, event := range events {
-		if event.Context.Tick > tick {
-			tick = event.Context.Tick
-		}
-	}
-	return tick
-}
-
 func eventStringSlice(values *[]string) []string {
 	if values == nil {
 		return nil
