+++ tests/functional/replay_contracts/replay_live_helpers_test.go
@@ -1,3 +1,5 @@
+//go:build functionallong
+
 package replay_contracts
 
 import (
