+++ tests/functional/replay_contracts/short_helpers_long_test.go
@@ -1,0 +1,35 @@
+//go:build functionallong
+
+package replay_contracts
+
+import (
+	factoryapi "github.com/portpowered/infinite-you/pkg/api/generated"
+	"github.com/portpowered/infinite-you/pkg/interfaces"
+)
+
+func replayEventCount(artifact *interfaces.ReplayArtifact, eventType factoryapi.FactoryEventType) int {
+	count := 0
+	for _, event := range artifact.Events {
+		if event.Type == eventType {
+			count++
+		}
+	}
+	return count
+}
+
+func factoryRelationsValue(value *[]factoryapi.Relation) []factoryapi.Relation {
+	if value == nil {
+		return nil
+	}
+	return *value
+}
+
+func lastFactoryEventTick(events []factoryapi.FactoryEvent) int {
+	tick := 0
+	for _, event := range events {
+		if event.Context.Tick > tick {
+			tick = event.Context.Tick
+		}
+	}
+	return tick
+}
