+++ tests/functional/smoke/service_pipeline_config_helpers_test.go
@@ -1,32 +1,5 @@
 package smoke
 
-func simpleServicePipelineConfig() map[string]any {
-	return map[string]any{
-		"workTypes": []map[string]any{
-			{
-				"name": "task",
-				"states": []map[string]string{
-					{"name": "init", "type": "INITIAL"},
-					{"name": "complete", "type": "TERMINAL"},
-					{"name": "failed", "type": "FAILED"},
-				},
-			},
-		},
-		"workers": []map[string]string{
-			{"name": "worker-a"},
-		},
-		"workstations": []map[string]any{
-			{
-				"name":      "process",
-				"worker":    "worker-a",
-				"inputs":    []map[string]string{{"workType": "task", "state": "init"}},
-				"outputs":   []map[string]string{{"workType": "task", "state": "complete"}},
-				"onFailure": []map[string]string{{"workType": "task", "state": "failed"}},
-			},
-		},
-	}
-}
-
 func twoStageServicePipelineConfig() map[string]any {
 	return map[string]any{
 		"workTypes": []map[string]any{
